+++ lib/src/models/effect.g.dart
@@ -12,6 +12,8 @@ extension EffectCopyWith on Effect {
     Color? color,
     Vector2D? offset,
     num? radius,
+    bool? showShadowBehindNode,
+    num? spread,
     EffectType? type,
     bool? visible,
   }) {
@@ -20,6 +22,8 @@ extension EffectCopyWith on Effect {
       color: color ?? this.color,
       offset: offset ?? this.offset,
       radius: radius ?? this.radius,
+      showShadowBehindNode: showShadowBehindNode ?? this.showShadowBehindNode,
+      spread: spread ?? this.spread,
       type: type ?? this.type,
       visible: visible ?? this.visible,
     );
@@ -35,6 +39,7 @@ Effect _$EffectFromJson(Map<String, dynamic> json) {
     type: _$enumDecodeNullable(_$EffectTypeEnumMap, json['type']),
     visible: json['visible'] as bool? ?? true,
     radius: json['radius'] as num?,
+    spread: json['spread'] as num?,
     color: json['color'] == null
         ? null
         : Color.fromJson(json['color'] as Map<String, dynamic>),
@@ -42,6 +47,7 @@ Effect _$EffectFromJson(Map<String, dynamic> json) {
     offset: json['offset'] == null
         ? null
         : Vector2D.fromJson(json['offset'] as Map<String, dynamic>),
+    showShadowBehindNode: json['showShadowBehindNode'] as bool?,
   );
 }
 
@@ -49,9 +55,11 @@ Map<String, dynamic> _$EffectToJson(Effect instance) => <String, dynamic>{
       'type': _$EffectTypeEnumMap[instance.type],
       'visible': instance.visible,
       'radius': instance.radius,
+      'spread': instance.spread,
       'color': instance.color,
       'blendMode': _$BlendModeEnumMap[instance.blendMode],
       'offset': instance.offset,
+      'showShadowBehindNode': instance.showShadowBehindNode,
     };
 
 K _$enumDecode<K, V>(