+++ lib/src/models/frame.dart
@@ -1,9 +1,5 @@
 import 'package:figma/src/converters/converters.dart';
-import 'package:figma/src/models/primary_axis_align_items.dart';
-import 'package:figma/src/models/counter_axis_align_items.dart';
 import 'package:figma/src/models/models.dart';
-import 'package:figma/src/models/primary_axis_sizing_mode.dart';
-import 'package:figma/src/models/counter_axis_sizing_mode.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:copy_with_extension/copy_with_extension.dart';
 
@@ -48,6 +44,11 @@ class Frame extends Node {
 
   /// Keep height and width constrained to same ratio
   final bool? preserveRatio;
+
+  /// This property is applicable only for direct children of auto-layout frames,
+  /// ignored otherwise. Determines whether a layer should stretch along the parent’s
+  /// primary axis. A `0` corresponds to a fixed size and `1` corresponds to stretch
+  final double? layoutGrow;
 
   /// Horizontal and vertical layout constraints for node
   final LayoutConstraint? constraints;
@@ -178,6 +179,7 @@ class Frame extends Node {
     this.exportSettings,
     this.blendMode,
     this.preserveRatio,
+    this.layoutGrow,
     this.constraints,
     this.layoutAlign,
     this.transitionNodeID,
@@ -227,6 +229,7 @@ class Frame extends Node {
         exportSettings,
         blendMode,
         preserveRatio,
+        layoutGrow,
         constraints,
         layoutAlign,
         transitionNodeID,