+++ lib/src/models/instance.g.dart
@@ -10,11 +10,12 @@ extension InstanceCopyWith on Instance {
   Instance copyWith({
     SizeRectangle? absoluteBoundingBox,
     BlendMode? blendMode,
-    List<Node>? children,
+    List<Node?>? children,
     bool? clipsContent,
     String? componentId,
     LayoutConstraint? constraints,
     double? cornerRadius,
+    CounterAxisAlignItems? counterAxisAlignItems,
     CounterAxisSizingMode? counterAxisSizingMode,
     List<Effect>? effects,
     List<ExportSetting>? exportSettings,
@@ -26,13 +27,20 @@ extension InstanceCopyWith on Instance {
     num? itemSpacing,
     LayoutAlign? layoutAlign,
     List<LayoutGrid>? layoutGrids,
+    double? layoutGrow,
     LayoutMode? layoutMode,
     bool? locked,
     String? name,
     double? opacity,
     OverflowDirection? overflowDirection,
+    double? paddingBottom,
+    double? paddingLeft,
+    double? paddingRight,
+    double? paddingTop,
     dynamic? pluginData,
     bool? preserveRatio,
+    PrimaryAxisAlignItems? primaryAxisAlignItems,
+    PrimaryAxisSizingMode? primaryAxisSizingMode,
     List<num>? rectangleCornerRadii,
     List<List<num>>? relativeTransform,
     dynamic? sharedPluginData,
@@ -54,6 +62,8 @@ extension InstanceCopyWith on Instance {
       componentId: componentId ?? this.componentId,
       constraints: constraints ?? this.constraints,
       cornerRadius: cornerRadius ?? this.cornerRadius,
+      counterAxisAlignItems:
+          counterAxisAlignItems ?? this.counterAxisAlignItems,
       counterAxisSizingMode:
           counterAxisSizingMode ?? this.counterAxisSizingMode,
       effects: effects ?? this.effects,
@@ -66,13 +76,22 @@ extension InstanceCopyWith on Instance {
       itemSpacing: itemSpacing ?? this.itemSpacing,
       layoutAlign: layoutAlign ?? this.layoutAlign,
       layoutGrids: layoutGrids ?? this.layoutGrids,
+      layoutGrow: layoutGrow ?? this.layoutGrow,
       layoutMode: layoutMode ?? this.layoutMode,
       locked: locked ?? this.locked,
       name: name ?? this.name,
       opacity: opacity ?? this.opacity,
       overflowDirection: overflowDirection ?? this.overflowDirection,
+      paddingBottom: paddingBottom ?? this.paddingBottom,
+      paddingLeft: paddingLeft ?? this.paddingLeft,
+      paddingRight: paddingRight ?? this.paddingRight,
+      paddingTop: paddingTop ?? this.paddingTop,
       pluginData: pluginData ?? this.pluginData,
       preserveRatio: preserveRatio ?? this.preserveRatio,
+      primaryAxisAlignItems:
+          primaryAxisAlignItems ?? this.primaryAxisAlignItems,
+      primaryAxisSizingMode:
+          primaryAxisSizingMode ?? this.primaryAxisSizingMode,
       rectangleCornerRadii: rectangleCornerRadii ?? this.rectangleCornerRadii,
       relativeTransform: relativeTransform ?? this.relativeTransform,
       sharedPluginData: sharedPluginData ?? this.sharedPluginData,
@@ -146,8 +165,18 @@ Instance _$InstanceFromJson(Map<String, dynamic> json) {
         .toList(),
     clipsContent: json['clipsContent'] as bool?,
     layoutMode: _$enumDecodeNullable(_$LayoutModeEnumMap, json['layoutMode']),
+    primaryAxisSizingMode: _$enumDecodeNullable(
+        _$PrimaryAxisSizingModeEnumMap, json['primaryAxisSizingMode']),
     counterAxisSizingMode: _$enumDecodeNullable(
         _$CounterAxisSizingModeEnumMap, json['counterAxisSizingMode']),
+    primaryAxisAlignItems: _$enumDecodeNullable(
+        _$PrimaryAxisAlignItemsEnumMap, json['primaryAxisAlignItems']),
+    counterAxisAlignItems: _$enumDecodeNullable(
+        _$CounterAxisAlignItemsEnumMap, json['counterAxisAlignItems']),
+    paddingLeft: (json['paddingLeft'] as num?)?.toDouble(),
+    paddingTop: (json['paddingTop'] as num?)?.toDouble(),
+    paddingRight: (json['paddingRight'] as num?)?.toDouble(),
+    paddingBottom: (json['paddingBottom'] as num?)?.toDouble(),
     horizontalPadding: json['horizontalPadding'] as num?,
     verticalPadding: json['verticalPadding'] as num?,
     itemSpacing: json['itemSpacing'] as num?,
@@ -161,6 +190,7 @@ Instance _$InstanceFromJson(Map<String, dynamic> json) {
         .toList(),
     isMask: json['isMask'] as bool?,
     isMaskOutline: json['isMaskOutline'] as bool?,
+    layoutGrow: (json['layoutGrow'] as num?)?.toDouble(),
   );
 }
 
@@ -182,8 +212,17 @@ Map<String, dynamic> _$InstanceToJson(Instance instance) => <String, dynamic>{
       'exportSettings': instance.exportSettings,
       'blendMode': _$BlendModeEnumMap[instance.blendMode],
       'preserveRatio': instance.preserveRatio,
+      'layoutGrow': instance.layoutGrow,
       'constraints': instance.constraints,
       'layoutAlign': _$LayoutAlignEnumMap[instance.layoutAlign],
+      'primaryAxisAlignItems':
+          _$PrimaryAxisAlignItemsEnumMap[instance.primaryAxisAlignItems],
+      'counterAxisAlignItems':
+          _$CounterAxisAlignItemsEnumMap[instance.counterAxisAlignItems],
+      'paddingLeft': instance.paddingLeft,
+      'paddingTop': instance.paddingTop,
+      'paddingRight': instance.paddingRight,
+      'paddingBottom': instance.paddingBottom,
       'transitionNodeID': instance.transitionNodeID,
       'transitionDuration': instance.transitionDuration,
       'transitionEasing': _$EasingTypeEnumMap[instance.transitionEasing],
@@ -195,6 +234,8 @@ Map<String, dynamic> _$InstanceToJson(Instance instance) => <String, dynamic>{
       'layoutMode': _$LayoutModeEnumMap[instance.layoutMode],
       'counterAxisSizingMode':
           _$CounterAxisSizingModeEnumMap[instance.counterAxisSizingMode],
+      'primaryAxisSizingMode':
+          _$PrimaryAxisSizingModeEnumMap[instance.primaryAxisSizingMode],
       'horizontalPadding': instance.horizontalPadding,
       'verticalPadding': instance.verticalPadding,
       'itemSpacing': instance.itemSpacing,
@@ -293,9 +334,27 @@ const _$LayoutModeEnumMap = {
   LayoutMode.vertical: 'VERTICAL',
 };
 
+const _$PrimaryAxisSizingModeEnumMap = {
+  PrimaryAxisSizingMode.fixed: 'FIXED',
+  PrimaryAxisSizingMode.auto: 'AUTO',
+};
+
 const _$CounterAxisSizingModeEnumMap = {
   CounterAxisSizingMode.fixed: 'FIXED',
   CounterAxisSizingMode.auto: 'AUTO',
+};
+
+const _$PrimaryAxisAlignItemsEnumMap = {
+  PrimaryAxisAlignItems.min: 'MIN',
+  PrimaryAxisAlignItems.center: 'CENTER',
+  PrimaryAxisAlignItems.max: 'MAX',
+  PrimaryAxisAlignItems.spaceBetween: 'SPACE_BETWEEN',
+};
+
+const _$CounterAxisAlignItemsEnumMap = {
+  CounterAxisAlignItems.min: 'MIN',
+  CounterAxisAlignItems.center: 'CENTER',
+  CounterAxisAlignItems.max: 'MAX',
 };
 
 const _$OverflowDirectionEnumMap = {